feat(suite-native): introduce IconWithSpinner
