+++ suite-native/module-trading/src/components/history/TradeDetailSheet/TradeDetailHeader.tsx
@@ -5,9 +5,8 @@ import {
     isFinalStatus,
     selectTradingTradeByOrderId,
 } from '@suite-common/trading';
-import { Box, CircularSpinner, VStack } from '@suite-native/atoms';
-import { Icon } from '@suite-native/icons';
-import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
+import { VStack } from '@suite-native/atoms';
+import { IconWithSpinner } from '@suite-native/trading-atoms';
 
 import { TradeDetailAlert } from './TradeDetailAlert';
 import { type TradeStatusStep, getTradeStatusStep } from '../../../utils/general/utils';
@@ -18,16 +17,7 @@ type TradeDetailHeaderProps = {
     onOpenedBrowser: () => void;
 };
 
-const iconWrapperStyle = prepareNativeStyle(utils => ({
-    backgroundColor: utils.colors.surfaceFillRaised,
-    borderRadius: utils.borders.radii.round,
-    padding: utils.spacings.sp12,
-    alignItems: 'center',
-    justifyContent: 'center',
-}));
-
 export const TradeDetailHeader = ({ orderId, onOpenedBrowser }: TradeDetailHeaderProps) => {
-    const { applyStyle, utils } = useNativeStyles();
     const trade = useSelector((state: TradingRootState) =>
         selectTradingTradeByOrderId(state, orderId),
     );
@@ -42,16 +32,7 @@ export const TradeDetailHeader = ({ orderId, onOpenedBrowser }: TradeDetailHeade
     if ((['success', 'pending'] as TradeStatusStep[]).includes(statusStep)) {
         return (
             <VStack spacing="sp16" alignItems="center" justifyContent="center">
-                <Box style={applyStyle(iconWrapperStyle)}>
-                    <Icon name="arrowsLeftRight" size="extraLarge" />
-                    {isInProgress && (
-                        <CircularSpinner
-                            size={utils.spacings.sp56}
-                            color="legacyBackgroundAlertYellowBold"
-                            width={3}
-                        />
-                    )}
-                </Box>
+                <IconWithSpinner iconName="arrowsLeftRight" isInProgress={isInProgress} />
                 <TradeStatusBadge status={trade.data.status} />
             </VStack>
         );
