+++ suite-native/trading-atoms/src/components/IconWithSpinner.tsx
@@ -1,0 +1,33 @@
+import { Box, CircularSpinner } from '@suite-native/atoms';
+import { Icon, type IconName } from '@suite-native/icons';
+import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
+
+export type IconWithSpinnerProps = {
+    isInProgress?: boolean;
+    iconName: IconName;
+};
+
+const iconWrapperStyle = prepareNativeStyle(utils => ({
+    backgroundColor: utils.colors.backgroundSurfaceElevation1,
+    borderRadius: utils.borders.radii.round,
+    padding: utils.spacings.sp12,
+    alignItems: 'center',
+    justifyContent: 'center',
+}));
+
+export const IconWithSpinner = ({ iconName, isInProgress = true }: IconWithSpinnerProps) => {
+    const { applyStyle, utils } = useNativeStyles();
+
+    return (
+        <Box style={applyStyle(iconWrapperStyle)}>
+            <Icon name={iconName} size="extraLarge" />
+            {isInProgress && (
+                <CircularSpinner
+                    size={utils.spacings.sp56}
+                    color="backgroundAlertYellowBold"
+                    width={3}
+                />
+            )}
+        </Box>
+    );
+};
